+++ src/components/atoms/containers/index.ts
@@ -1,0 +1,6 @@
+export { AsideContainer } from "./AsideContainer";
+export { BottomAreaContainer } from "./BottomAreaContainer";
+export { ContainerWithOrder } from "./ContainerWithOrder";
+export { MainContainer } from "./MainContainer";
+export { MiddleAreaContainer } from "./MiddleAreaContainer";
+export { SideSectionContainer } from "./SideSectionContainer";
